+++ Mods ChoGGi/Fix Tourist Overview Softlock/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2434359529",
 	"pops_any_uuid", "2dffb277-a8f8-438b-b9a8-aa78da0333f9",
 	"lua_revision", 1001514, -- Tito
-	"version", 2,
+	"version", 3,
 	"version_major", 0,
-	"version_minor", 2,
+	"version_minor", 3,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -14,14 +14,13 @@ return PlaceObj("ModDef", {
 	},
 	"has_options", true,
 	"TagOther", true,
-	"description", [[Fix the softlock from Tourism Update- Hotfix 3.
+	"description", [[
+OBSOLETE: Fixed in https://forum.paradoxplaza.com/forum/threads/small-hotfix.1470386/
+This mod will do nothing on revision 1001586 or later
 
 
-customSupplyRocket.lua > "OnPressParam", "UIOpenTouristOverview",
 
-SupplyRocket.lua > SupplyRocket:UIOpenTouristOverview(reward_info)
-instead of getting the reward_info table it gets 1
-
+Fixes the softlock from Tourism Update- Hotfix 3.
 
 Compatible with Automated Tourism
 ]],